+++ vector/src/main/java/im/vector/riotx/core/error/ErrorFormatter.kt
@@ -17,6 +17,7 @@
 package im.vector.riotx.core.error
 
 import im.vector.matrix.android.api.failure.Failure
+import im.vector.matrix.android.api.failure.MatrixError
 import im.vector.riotx.R
 import im.vector.riotx.core.resources.StringProvider
 import javax.inject.Inject
@@ -34,8 +35,13 @@ class ErrorFormatter @Inject constructor(val stringProvider: StringProvider) {
             null                         -> null
             is Failure.NetworkConnection -> stringProvider.getString(R.string.error_no_network)
             is Failure.ServerError       -> {
-                throwable.error.message.takeIf { it.isNotEmpty() }
-                        ?: throwable.error.code.takeIf { it.isNotEmpty() }
+                if (throwable.error.code == MatrixError.M_CONSENT_NOT_GIVEN) {
+                    // Special case for terms and conditions
+                    stringProvider.getString(R.string.error_terms_not_accepted)
+                } else {
+                    throwable.error.message.takeIf { it.isNotEmpty() }
+                            ?: throwable.error.code.takeIf { it.isNotEmpty() }
+                }
             }
             else                         -> throwable.localizedMessage
         }