+++ vector/src/main/java/im/vector/riotx/core/platform/VectorBaseActivity.kt
@@ -36,11 +36,14 @@ import butterknife.Unbinder
 import com.airbnb.mvrx.BaseMvRxActivity
 import com.bumptech.glide.util.Util
 import com.google.android.material.snackbar.Snackbar
+import im.vector.matrix.android.api.failure.ConsentNotGivenError
 import im.vector.riotx.BuildConfig
 import im.vector.riotx.R
 import im.vector.riotx.core.di.*
+import im.vector.riotx.core.dialogs.DialogLocker
 import im.vector.riotx.core.utils.toast
 import im.vector.riotx.features.configuration.VectorConfiguration
+import im.vector.riotx.features.consent.ConsentNotGivenHelper
 import im.vector.riotx.features.navigation.Navigator
 import im.vector.riotx.features.rageshake.BugReportActivity
 import im.vector.riotx.features.rageshake.BugReporter
@@ -50,6 +53,9 @@ import im.vector.riotx.features.themes.ThemeUtils
 import im.vector.riotx.receivers.DebugReceiver
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
+import org.greenrobot.eventbus.EventBus
+import org.greenrobot.eventbus.Subscribe
+import org.greenrobot.eventbus.ThreadMode
 import timber.log.Timber
 import kotlin.system.measureTimeMillis
 
@@ -392,6 +398,31 @@ abstract class VectorBaseActivity : BaseMvRxActivity(), HasScreenInjector {
     }
 
     /* ==========================================================================================
+     * User Consent
+     * ========================================================================================== */
+
+    private val consentNotGivenHelper by lazy {
+        ConsentNotGivenHelper(this, DialogLocker(savedInstanceState))
+                .apply { restorables.add(this) }
+    }
+
+    override fun onStart() {
+        super.onStart()
+        EventBus.getDefault().register(this)
+    }
+
+    override fun onStop() {
+        super.onStop()
+        EventBus.getDefault().unregister(this)
+    }
+
+    @Subscribe(threadMode = ThreadMode.MAIN)
+    fun onConsentNotGivenError(consentNotGivenError: ConsentNotGivenError) {
+        consentNotGivenHelper.displayDialog(consentNotGivenError.consentUri,
+                screenComponent.session().sessionParams.homeServerConnectionConfig.homeServerUri.host ?: "")
+    }
+
+    /* ==========================================================================================
      * Temporary method
      * ========================================================================================== */
 
@@ -402,5 +433,4 @@ abstract class VectorBaseActivity : BaseMvRxActivity(), HasScreenInjector {
             toast(getString(R.string.not_implemented))
         }
     }
-
 }